feat: spreadsheetification of lottery export
