+++ api/package.json
@@ -60,6 +60,7 @@
     "compression": "^1.7.4",
     "cookie-parser": "~1.4.6",
     "dayjs": "~1.11.9",
+    "exceljs": "^4.4.0",
     "handlebars": "~4.7.8",
     "jsonwebtoken": "~9.0.1",
     "lodash": "~4.17.21",
@@ -80,6 +81,7 @@
     "@nestjs/schematics": "^10.1.1",
     "@nestjs/testing": "^10.3.2",
     "@types/compression": "^1.7.5",
+    "@types/exceljs": "1.3.0",
     "@types/express": "~4.17.17",
     "@types/jest": "~29.5.3",
     "@types/node": "^18.7.14",
