+++ shared-helpers/src/types/backend-swagger.ts
@@ -1444,35 +1444,6 @@ export class ApplicationsService {
     })
   }
   /**
-   * Get applications lottery results
-   */
-  lotteryResults(
-    params: {
-      /**  */
-      listingId: string
-      /**  */
-      includeDemographics?: boolean
-      /**  */
-      timeZone?: string
-    } = {} as any,
-    options: IRequestOptions = {}
-  ): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let url = basePath + "/applications/getLotteryResults"
-
-      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
-      configs.params = {
-        listingId: params["listingId"],
-        includeDemographics: params["includeDemographics"],
-        timeZone: params["timeZone"],
-      }
-
-      /** 适配ios13，get请求不允许带body */
-
-      axios(configs, resolve, reject)
-    })
-  }
-  /**
    * Get applications as csv
    */
   listAsCsv(
@@ -2172,6 +2143,35 @@ export class LotteryService {
       let data = params.body
 
       configs.data = data
+
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Get applications lottery results
+   */
+  lotteryResults(
+    params: {
+      /**  */
+      listingId: string
+      /**  */
+      includeDemographics?: boolean
+      /**  */
+      timeZone?: string
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<any> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/lottery/getLotteryResults"
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+      configs.params = {
+        listingId: params["listingId"],
+        includeDemographics: params["includeDemographics"],
+        timeZone: params["timeZone"],
+      }
+
+      /** 适配ios13，get请求不允许带body */
 
       axios(configs, resolve, reject)
     })
